requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
requests==2.31.0
python-dotenv==1.0.1
python-multipart==0.0.9
